% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_nas.R
\name{assign_nas}
\alias{assign_nas}
\title{Cleaning NAs}
\usage{
assign_nas(data, column, additional_nas = c(91, 92))
}
\arguments{
\item{data}{A dataset}

\item{column}{A column within a dataset}

\item{additional_nas}{A vector of additional values to reassign to NAs ("91" and "92" by default)}
}
\value{
The original column with certain values reassigned to NA
}
\description{
A function to reassign additional values as NAs.
}
