\name{citbir-package}
\alias{citbir-package}
\alias{citbir}
\docType{package}
\title{
\packageTitle{citbir}
}
\description{
\packageDescription{citbir}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{citbir}

Maintainer: \packageMaintainer{citbir}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = citbir)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{citbir}}
%\section{Documentation Index}{\packageIndices{citbir}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
