% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_positive.R
\name{true_positive}
\alias{true_positive}
\title{Finding Type I and II Errors}
\usage{
true_positive(truths, fitted)
}
\arguments{
\item{truths}{A vector of real, observed values.}

\item{fitted}{A vector of predicted/fitted values.}
}
\value{
An integer.
}
\description{
A family of functions for deriving the number of true positives, true negatives, false positives, and false negatives from two vectors.
}
